feat(sdk): move the grammar-shape access types into the SDK
GrammarShape, ResolvedCall, and GrammarModelShape relocate to
packages/dxkit-sdk/src/grammar.ts — the read contract rung-4 plugins
get grammar access through. The per-grammar rows, factories, and
dispatch stay internal in src/ast/; both modules re-export the types
so no consumer changes an import.
